Move examples files to examples root
